Split EditableActionTable into several files
